pyproject.toml
[project]
name = "coding-agent-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai==1.12.1",
    "openai>=2.21.0",
    "python-dotenv==1.1.0",
]

[tool.uv.workspace]
members = [
    "calculator",
]
